requirements.txt
# Core tools (fork.py, promote.py, registry.py)
pyyaml>=6.0

# Evaluation harness
human-eval>=1.0

# Model inference (real model loading in humaneval.py)
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
peft>=0.10.0
